refactor: rename "StakingSliderAnnotation" to avoid confusion with "ObjectionAnnotation"
StakingSliderAnnotation is not actually an Annotation component, it's more of a label that sits above the staking slider.
