--- src/components/common/ColonyActions/ActionDetailsPage/DefaultMotion/MotionPhaseWidget/StakingWidget/StakingInput/StakingSliderAnnotation.css.d.ts
@@ -1,13 +1,0 @@
-declare namespace StakingSliderAnnotationCssNamespace {
-  export interface IStakingSliderAnnotationCss {
-    minStakeAmountContainer: string;
-    tooltip: string;
-  }
-}
-
-declare const StakingSliderAnnotationCssModule: StakingSliderAnnotationCssNamespace.IStakingSliderAnnotationCss & {
-  /** WARNING: Only available when `css-loader` is used without `style-loader` or `mini-css-extract-plugin` */
-  locals: StakingSliderAnnotationCssNamespace.IStakingSliderAnnotationCss;
-};
-
-export = StakingSliderAnnotationCssModule;
