+++ src/components/common/ColonyActions/ActionDetailsPage/DefaultMotion/MotionPhaseWidget/StakingWidget/StakingInput/StakingSliderMessages/index.ts
@@ -1,1 +1,4 @@
-export { default as RequiredStakeMessage } from './RequiredStakeMessage';
+export {
+  default as RequiredStakeMessage,
+  RequiredStakeMessageProps,
+} from './RequiredStakeMessage';
