+++ src/components/common/ColonyActions/ActionDetailsPage/DefaultMotion/MotionPhaseWidget/StakingWidget/StakingInput/index.ts
@@ -1,6 +1,6 @@
 export { default, SLIDER_AMOUNT_KEY } from './StakingInput';
 export { default as StakingControls } from './StakingControls';
-export { default as StakingSliderAnnotation } from './StakingSliderAnnotation';
+export { default as StakingSliderLabel } from './StakingSliderLabel';
 export { default as StakingSliderDescription } from './StakingSliderDescription';
 export { default as StakingWidgetSlider } from './StakingWidgetSlider';
 export { default as StakingSlider } from './StakingSlider';
